Update README.md
The "Image Gallery" project is a sophisticated web application designed to showcase a collection of images in an organized and visually engaging manner. Leveraging the power of HTML, CSS, and JavaScript, this project highlights the seamless integration of these technologies to create a dynamic and responsive user experience. Developed using Visual Studio Code, this project exemplifies best practices in modern web development, making it an ideal learning resource for both beginners and experienced developers.